partially rearrange modules
